remove swizzling
