--- FantasmoSDK/Classes/FantasmoLoader.h
@@ -1,16 +1,0 @@
-//
-//  NSObject+Loader.h
-//  FantasmoSDK
-//
-//  Copyright © 2020 Fantasmo. All rights reserved.
-//
-
-#import <Foundation/Foundation.h>
-
-NS_ASSUME_NONNULL_BEGIN
-
-@interface FantasmoLoader: NSObject
-
-@end
-
-NS_ASSUME_NONNULL_END
